# Sbromberger/remove bitvecs2