--- a/src/digraph/cycles/hadwick-james.jl
+++ b/src/digraph/cycles/hadwick-james.jl
@@ -15,7 +15,7 @@ function simplecycles_hadwick_james end
     for i in vertices(g)
         push!(B, Vector{T}())
     end
-    blocked = falses(nvg)
+    blocked = zeros(Bool, nvg)
     stack = Vector{T}()
     cycles = Vector{Vector{T}}()
     for v in vertices(g)
@@ -46,7 +46,7 @@ resetblocked!(blocked) = fill!(blocked, false)
 Find circuits in `g` recursively starting from v1.
 """
 function circuit_recursive! end
-@traitfn function circuit_recursive!{T<:Integer}(g::::IsDirected, v1::T, v2::T, blocked::BitArray, B::Vector{Vector{T}}, stack::Vector{T}, cycles::Vector{Vector{T}})
+@traitfn function circuit_recursive!{T<:Integer}(g::::IsDirected, v1::T, v2::T, blocked::AbstractVector, B::Vector{Vector{T}}, stack::Vector{T}, cycles::Vector{Vector{T}})
     f = false
     push!(stack, v2)
     blocked[v2] = true
@@ -80,7 +80,7 @@ end
 
 Unblock the value `v` from the `blocked` list and remove from `B`.
 """
-function unblock!(v::T, blocked::BitArray, B::Vector{Vector{T}}) where T
+function unblock!(v::T, blocked::AbstractVector, B::Vector{Vector{T}}) where T
     blocked[v] = false
     wPos = 1
     Bv = B[v]