Remove bitvecs for two functions. Benchmarks below.