@sbromberger 2017-08-17
```
julia> g = PathDiGraph(10_000)
{10000, 9999} directed simple Int64 graph

# Current master
julia> @benchmark LightGraphs.simplecycles_hadwick_james($g)
BenchmarkTools.Trial:
  memory estimate:  1.74 MiB
  allocs estimate:  20035
  --------------
  minimum time:     5.581 s (0.00% GC)
  median time:      5.581 s (0.00% GC)
  mean time:        5.581 s (0.00% GC)
  maximum time:     5.581 s (0.00% GC)
  --------------
  samples:          1
  evals/sample:     1

# This PR
julia> @benchmark LightGraphs.simplecycles_hadwick_james2($g)
BenchmarkTools.Trial:
  memory estimate:  1.75 MiB
  allocs estimate:  20032
  --------------
  minimum time:     5.255 s (0.00% GC)
  median time:      5.255 s (0.00% GC)
  mean time:        5.255 s (0.00% GC)
  maximum time:     5.255 s (0.00% GC)
  --------------
  samples:          1
  evals/sample:     1
```

```
julia> g = CompleteBipartiteGraph(1000,1000)
{2000, 1000000} undirected simple Int64 graph

# Current master
julia> @benchmark bipartite_map(g)
BenchmarkTools.Trial:
  memory estimate:  121.50 KiB
  allocs estimate:  55
  --------------
  minimum time:     12.099 ms (0.00% GC)
  median time:      13.271 ms (0.00% GC)
  mean time:        13.532 ms (0.07% GC)
  maximum time:     21.636 ms (6.77% GC)
  --------------
  samples:          370
  evals/sample:     1

# This PR
julia> @benchmark bipartite_map($g)
BenchmarkTools.Trial:
  memory estimate:  124.84 KiB
  allocs estimate:  49
  --------------
  minimum time:     6.181 ms (0.00% GC)
  median time:      6.808 ms (0.00% GC)
  mean time:        6.937 ms (0.07% GC)
  maximum time:     11.459 ms (0.00% GC)
  --------------
  samples:          720
  evals/sample:     1
```