@sbromberger 2017-08-22
I will be merging this by end of day tomorrow unless there are objections.